Added a CardText component
This component renders a block that is meant to go inside the Card
component (#62) with the following properties:

- Title
- URL
- Short description
- Publisher
- Author

A couple of unit tests have been added to test the CardText component,
and the CRA-provided default test for App marked as skipped as it no
longer passes and is not needed just yet.

I have also changed the comment line in App.tsx from <!-- --> to {/* */}
as the former produces a TypeScript error and doesn't let the test suite run.

Resolves #64.