Add tests to layout components
- Added Prettier config specified in Pocket#86 so that the newly created
tests already use it.

- Normalised component names/files/folders, i.e PocketCard component
is now Card in Card.tsx, Button component is now in Button.tsx and not
ButtonComponent.tsx, etc.

- Moved AddStory component into its own folder. Replaced h1 element styled
in a separate .css file with a Typography element.

- Added basic state management and onChange handlers to LoginForm
and AddStory input fields.

- Button: text is now passed as children to align it with the MUI Button.

- Added tests for Button, AddStory, LoginForm components.

Note that the bulk of the rewritten code is Prettier adjusting tab width,
adding commas, etc. My changes to existing components are minor - mainly
with the goal of making them testable.

Closes Pocket#85. Also closes Pocket#86 (Prettier config).