Upgrade to TypeScript 4.1.3
In #89, it says that the updated Create React App has TypeScript 4.1.
That was not quite correct: it is compatible with 4.1 but the actual dependency
is ^3.2, so the 3.7.5 version installed when this project was started was enough.

Updated TypeScript to the latest version (4.1.3) as well as the @types/react
package that needed to be updated, too. There are some new features and it also
catches extra errors - #106 has a fixup commit for a functional component.