Update the data generation script for local API
Delete db.json file before overwriting it with new data as truncating it
is not enough if schema has been updated - the frontend app shows a blank
screen even after following the steps in the troubleshooting section.

Improve logging messages to console, add a helpful hint with a reminder
which command to run next.